Integrated SimulinkModelGenerator to translate NetworkUI nodes to simulink model output

+++ OpenSTSM/OpenSTSM/Converters/ListOfDecimalToStringConverter.cs
@@ -5,12 +5,12 @@ using OpenSTSM.Extensions;
 
 namespace OpenSTSM.Converters
 {
-    [ValueConversion(typeof(List<decimal>), typeof(string))]
+    [ValueConversion(typeof(List<double>), typeof(string))]
     public class ListOfDecimalToStringConverter : IValueConverter
     {
         public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
         {            
-            var result = string.Join(" ", ((List<decimal>)value).ToArray());
+            var result = string.Join(" ", ((List<double>)value).ToArray());
             return $"[{result}]";
         }
 
@@ -19,9 +19,9 @@ namespace OpenSTSM.Converters
             string input = (string)value;
             input = string.Concat(input.RemoveChars('[', ']'));
             input = input.TrimStart().TrimEnd();
-            List<decimal> coefficients = new List<decimal>();
+            List<double> coefficients = new List<double>();
             var coefficients_string = input.SplitStringOnChar(' ');
-            coefficients_string.ForEach(c => coefficients.Add(decimal.Parse(c)));
+            coefficients_string.ForEach(c => coefficients.Add(double.Parse(c)));
             return coefficients;
         }
     }
